Add state machine for validating token sequences (#8)
Both the Encoder and Decoder need to validate whether
a given sequence of JSON tokens is valid or not.

We implement stateMachine, which is a push-down automaton with
methods to append each token kind. If it returns an error,
then that token is invalid as next in the sequence
(and the state machine remains unchanged). Otherwise, it is valid and
the state machine is updated to reflect the appending of that token.

The stateMachine methods are carefully written to be inlineable
for performance reasons.