fix errors in the NewEncoder example
ObjectStart and company were written backwards, which seems to have been
a mistake from the beginning. Perhaps from a very early on rename that
never made it to the git history.

Number also doesn't exist. I assume that should be Float now.

We should probably move this to be a real Example func, which would have
caught these mistakes via "go test". For now, just fix the mistakes.