Add RawValue.Clone and RawValue.String methods (#93)
Encoder.ReadValue warns that the returned value is only valid until
the next call. However, the user may want to hold on to the raw value,
in which case a RawValue.Clone method is useful.

JSON is a humanly readable format.
We should print RawValue as is, rather than as a slice of uint8.