Detect cyclic data structures (#95)
In v1 encoding/json, support for detecting and rejecting
cyclic data structures was added in:
  * https://golang.org/cl/187920
  * https://golang.org/cl/248358

In Go, cycles can only occur through pointers, slices, and maps,
which are the three data structures that we check for in v2.

Similar to the implementation in v1, we only check for cycles
when the depth is beyond a certain limit. In practice,
this limit is almost never triggered.

Cycle detection may experience false negatives since we cannot
detect cycles that occur across MarshalJSON calls because
we have no mechanism for passing the set of visited pointers.
This failure mode exists in v1 as well.