Implement omitempty struct tag feature (#110)
We define omitzero as omitting a field
if it is the zero value in Go.
We define omitempty as omitting a field
if it would have marshaled as an empty JSON value.

We define an "empty" JSON value as either
null, "", {}, or [].

The naive approach requires always marshaling
the value and unwriting it if it turns out
to be empty. As an optimization, we can prove
that a value would have serialized as an empty
JSON value if there no custom marshalers.