Manually check for duplicate names in Marshal and Unmarshal
By default, the Encoder and Decoder check that JSON objects do not
contain any duplicate names. This is more correct, but also a relative
expensive check.

However, since we are most often marshaling JSON objects to/from a
Go struct or Go map, we can use that unique property to manually perform
duplicate name checks in a more efficient way.
When unmarshaling into a Go struct,
we can use a bit set keyed by the Go struct field index.
When unmarshaling into a Go map,
we can use pre-existing presence within the map itself as a signal for duplicates.

As such, we add functionality for the arshal logic to specially disable
namespaces, knowing that it will manually check for duplicate names.
While this provides a performance improvement, it can leave the
Encoder or Decoder in an inconsistent state when errors are encountered.
When this occurs, we lock up the coder such that future methods
return errInvalidNamespace.

The performance gains of this optimization are notable:
	Testdata/GolangSource/Marshal/Concrete      154MB/s ± 3%    181MB/s ± 3%  +17.20%
	Testdata/GolangSource/Marshal/Interface    48.6MB/s ± 2%   54.2MB/s ± 2%  +11.48%
	Testdata/GolangSource/Unmarshal/Concrete    128MB/s ± 1%    157MB/s ± 1%  +22.97%
	Testdata/GolangSource/Unmarshal/Interface  46.7MB/s ± 1%   50.9MB/s ± 1%   +8.91%
	Testdata/SyntheaFhir/Marshal/Concrete      61.0MB/s ± 1%  102.2MB/s ± 1%  +67.49%
	Testdata/SyntheaFhir/Marshal/Interface      108MB/s ± 2%    121MB/s ± 1%  +12.44%
	Testdata/SyntheaFhir/Unmarshal/Concrete     205MB/s ± 1%    226MB/s ± 1%  +10.29%
	Testdata/SyntheaFhir/Unmarshal/Interface   85.9MB/s ± 1%   89.3MB/s ± 2%   +3.95%

However, it is still not as fast as disabling the check completely:
	Testdata/GolangSource/Marshal/Concrete      181MB/s ± 3%    189MB/s ± 2%   +4.62%
	Testdata/GolangSource/Marshal/Interface    54.2MB/s ± 2%   53.5MB/s ± 1%   -1.34%
	Testdata/GolangSource/Unmarshal/Concrete    157MB/s ± 1%    153MB/s ± 1%   -2.43%
	Testdata/GolangSource/Unmarshal/Interface  50.9MB/s ± 1%   51.0MB/s ± 1%     ~
	Testdata/SyntheaFhir/Marshal/Concrete       102MB/s ± 1%    108MB/s ± 1%   +5.43%
	Testdata/SyntheaFhir/Marshal/Interface      121MB/s ± 1%    119MB/s ± 2%   -1.90%
	Testdata/SyntheaFhir/Unmarshal/Concrete     226MB/s ± 1%    240MB/s ± 2%   +5.99%
	Testdata/SyntheaFhir/Unmarshal/Interface   89.3MB/s ± 2%   93.2MB/s ± 1%   +4.41%