Maintain a stack of all object names (#128)
For better error reporting, we would like to report exactly
where a problem occurred. However, this requires that the
state machine remember the sequence of object names.

A naive implementation pushes the current object name
onto a stack when visting a nested object, and
pops the name when leaving the nested object.
However, this incurs a significant performance cost
copying the object name repeatedly:

	Testdata/GolangSource/Encode/Token/Streaming-24        333MB/s ± 1%    246MB/s ± 0%      -26.04%  (p=0.000 n=9+9)
	Testdata/GolangSource/Decode/Token/Streaming-24        569MB/s ± 3%    394MB/s ± 1%      -30.72%  (p=0.000 n=10+10)
	Testdata/StarcraftSettings/Encode/Token/Streaming-24   572MB/s ± 1%    365MB/s ± 1%      -36.27%  (p=0.000 n=10+9)
	Testdata/StarcraftSettings/Decode/Token/Streaming-24   675MB/s ± 1%    424MB/s ± 0%      -37.27%  (p=0.000 n=10+8)
	Testdata/SyntheaFhir/Encode/Token/Streaming-24         922MB/s ± 1%    694MB/s ± 2%      -24.74%  (p=0.000 n=9+10)
	Testdata/SyntheaFhir/Decode/Token/Streaming-24         831MB/s ± 1%    708MB/s ± 1%      -14.79%  (p=0.000 n=10+10)

The implementation in this change is more clever where it
only stores the offset into the encodeBuffer or decodeBuffer
of where the name is located. The cost of tracking
the name only involves updating a single integer.
However, there is trickiness where we have to copy the names
to a local buffer whenever Decoder.fetch or Encoder.flush
is called since those mutate the buffer.
This implementation still carries a cost, but not as severe:

	Testdata/GolangSource/Encode/Token/Streaming-24        333MB/s ± 1%   316MB/s ± 3%    -4.96%  (p=0.000 n=9+10)
	Testdata/GolangSource/Decode/Token/Streaming-24        569MB/s ± 3%   511MB/s ± 1%   -10.13%  (p=0.000 n=10+10)
	Testdata/StarcraftSettings/Encode/Token/Streaming-24   572MB/s ± 1%   481MB/s ± 1%   -15.88%  (p=0.000 n=10+10)
	Testdata/StarcraftSettings/Decode/Token/Streaming-24   675MB/s ± 1%   581MB/s ± 1%   -14.03%  (p=0.000 n=10+10)
	Testdata/SyntheaFhir/Encode/Token/Streaming-24         922MB/s ± 1%   837MB/s ± 2%    -9.23%  (p=0.000 n=9+10)
	Testdata/SyntheaFhir/Decode/Token/Streaming-24         831MB/s ± 1%   779MB/s ± 2%    -6.35%  (p=0.000 n=10+10)

The performance cost is more notable when streaming since fetch/flush
forces the object names to be copied. The cost is most notable when
streaming StarcraftSettings dataset since the file is relatively short.
Thus, the overhead of allocating the stack dominates the cost.
When operating on fully buffered input/output, the cost is negligible.

This is a zero sum game.
Either we have good errors and pay the cost to track object names
or we avoid tracking object names and have poor errors.