Optimize string for multi-byte runes (#142)
Calling append(out, in[:runeSize]...) for every rune is expensive
since the overhead of calling runtime.memmove is high relative to
the work of actually copying at most 4 bytes.

Modify appendString to copy unescaped input in larger batches.
This undoes a previous optimization where we append each byte
as we process it in the pure ASCII case. The slows down the ASCII
case by ~1%, but speeds up multibyte runes by ~26%.

As a minor optimization, we track the position in the input as an offset,
rather by slicing off the processed prefix. This is slightly faster
since updating an offset only needs to modify a single register,
while slicing requires updating the entire string header
and a bounds check to ensure that the slice range is valid.

Performance:
	Testdata/GolangSource/Marshal/Concrete   5.81ms ± 3%  5.88ms ± 3%   +1.28%  (p=0.000 n=50+49)
	Testdata/StringUnicode/Marshal/Concrete  49.7µs ± 1%  36.8µs ± 1%  -25.98%  (p=0.000 n=110+106)
	Testdata/SyntheaFhir/Marshal/Concrete    12.9ms ± 3%  12.9ms ± 3%   +0.63%  (p=0.000 n=108+106)