Add stringCache for interning strings (#119)
This implements a cache for recently allocated strings.
It is functionally a lossy hash map with a fixed size.

The hash algorithm chosen is based on xxhash since the algorithm
relies on uint64 integer operations, which nicely matches the fact
that we already have two uint64s on hand. Other algorithms like
CRC, Adler, FNV, etc. operate on a byte-by-byte basis.
We skip the avalanche step of XXH64 since it costs 5 additional
machine instructions for negligible benefit for our application.

Performance:

	BenchmarkIntern/Best/Alloc      15629 ns/op   16000 B/op   1000 allocs/op
	BenchmarkIntern/Best/Cache       5855 ns/op      16 B/op      1 allocs/op
	BenchmarkIntern/Best/GoMap       5382 ns/op      16 B/op      1 allocs/op

	BenchmarkIntern/Repeat/Alloc    12769 ns/op    9600 B/op   1000 allocs/op
	BenchmarkIntern/Repeat/Cache     5955 ns/op      96 B/op     10 allocs/op
	BenchmarkIntern/Repeat/GoMap    16558 ns/op     678 B/op     11 allocs/op

	BenchmarkIntern/Synthea/Alloc  461717 ns/op  709849 B/op  24672 allocs/op
	BenchmarkIntern/Synthea/Cache  346423 ns/op  231208 B/op   6548 allocs/op
	BenchmarkIntern/Synthea/GoMap  909352 ns/op  431794 B/op   2520 allocs/op

	BenchmarkIntern/Twitter/Alloc  105988 ns/op  217592 B/op   4609 allocs/op
	BenchmarkIntern/Twitter/Cache  100348 ns/op  158832 B/op   2344 allocs/op
	BenchmarkIntern/Twitter/GoMap  290935 ns/op  259683 B/op   1546 allocs/op

	BenchmarkIntern/Worst/Alloc     15131 ns/op   16000 B/op   1000 allocs/op
	BenchmarkIntern/Worst/Cache     21617 ns/op   16000 B/op   1000 allocs/op
	BenchmarkIntern/Worst/GoMap    127348 ns/op  182483 B/op   1030 allocs/op

Alloc is the performance relying purely on the Go allocator,
Cache is the performance relying on stringCache, and
GoMap is the performance relying on a naive Go map.

In the worst case, the cache tends to run 1.4x slower than simply allocating.
However, in the case where there are many duplicated strings,
the cache can be 1.3x to 3x faster and reduce allocations significantly.

The string cache seems to be a superior solution compared to a naive Go map:
  * It avoids unbounded growth of the cache.
  * It has very fast lookups mostly independent of the string length N.
    The cache lookup cost is generally O(1) for a miss and O(N) for a hit.
    In contrast, lookups in a map are almost always O(N).

However, there are detriments as well:
  * The cache has a fixed size regardless of utilization, while a Go map
    occupies O(N) space, where N is the number of interned strings.
    This could be resolved by using a dynamically sized cache.
  * Cache utilization is poor. In order to keep lookups fast,
    we only check a single slot per hash. Due to the Birthday Paradox,
    there is a high probability that we evict another interned string.
    Inserting N unique strings into a cache of size N will not actually
    result in 100% utilization, but something closer to 63%.
    This can be improved through chaining, where we allow using up to K
    slots to store a string. While this improves utilization, it increases
    lookup costs by a factor of K.

Use of the cache provides performance benefits when unmarshaling:

	Testdata/CanadaGeometry/Unmarshal/Concrete  2.70ms ± 6%  2.61ms ± 5%   -3.14%  (p=0.000 n=50+47)
	Testdata/CitmCatalog/Unmarshal/Concrete     5.00ms ± 4%  5.03ms ± 3%   +0.64%  (p=0.024 n=49+50)
	Testdata/GolangSource/Unmarshal/Concrete    11.5ms ± 4%  11.1ms ± 5%   -3.95%  (p=0.000 n=50+48)
	Testdata/StringUnicode/Unmarshal/Concrete   65.1µs ± 3%  37.2µs ± 5%  -42.79%  (p=0.000 n=48+50)
	Testdata/SyntheaFhir/Unmarshal/Concrete     8.13ms ± 5%  7.54ms ± 3%   -7.22%  (p=0.000 n=49+48)
	Testdata/TwitterStatus/Unmarshal/Concrete   2.11ms ± 4%  1.90ms ± 3%  -10.02%  (p=0.000 n=50+47)

Since benchmarks repeatedly unmarshal the same dataset,
the use of a cache would unfairly benefit this situation.
The above results were obtained by artificially hindering the cache by zeroing
it out for every call to UnmarshalOptions.Unmarshal.
Thus, string interning only occurs for strings within a single invocation
of Unmarshal for benchmarking purposes. Even so, there is a notable benefit
to the use of string interning.