Always visit seen nodes when deriving JSON representable fields (#204)
In order for cancelation of fields with the same JSON name
at the same inlining depth, we must always visit a type
even if we have seen it before. To prevent infinite recursion,
prevent a type that we've seen before from being able to add
more types to the BFS queue.