Rename and move some examples (#217)
Rename examples to improve what declaration they are grouped under.
Move the OrderedObject example to a separate file.
This example references types declared outside the example body
that should be shown with the example. Placing this in a separate
file is a clear signal to the pkg.go.dev that this should be printed together.