Require that the format tag option comes last (#243)
If there is any tag option that is most likely to grow in complexity,
it is the `format` tag option.

Today we can have:

	struct { M map[string]int `json:",format:emitnull"` }

Let's suppose we hypothetically support for ordered marshaling of maps
with the `ordered` format flag. It would be tempting to write:

	struct { M map[string]int `json:",format:emitnull,ordered"` }

However, this presents a grammatical ambiguity where it is unclear
whether `ordered` belongs to `format` or the outer-level list of tag options.

To avoid ambiguities, we would have to do something like the following:

	struct { M map[string]int `json:",format:'emitnull,ordered'"` }
	struct { M map[string]int `json:",format:{emitnull,ordered}"` }

Restrict the `format` option to be the last one so that static tools
can detect misuse in the future (if we ever add composite `format` flags).
This change does NOT add support for composite tag options.