Rely on new reflect.Value.MapRange semantics in Go 1.19 (#248)
The allocating part of MapRange has been placed in a function
that is inlineable so that the MapIter value can be stack allocated
in most calling usages. There is no need to pool the MapIter value.

Performance:

	name                     old time/op    new time/op    delta
	Marshal/Map/ManyEmpty    2.02µs ± 3%    1.92µs ± 5%   -5.07%  (p=0.000 n=20+20)
	Marshal/Map/OneLarge     2.13µs ± 1%    2.26µs ± 1%   +5.80%  (p=0.000 n=20+18)
	Marshal/Map/ManySmall    7.96µs ± 1%    7.19µs ± 2%   -9.63%  (p=0.000 n=19+20)

	name                       old speed      new speed    delta
	Marshal/Map/ManyEmpty   149MB/s ± 3%   157MB/s ± 5%   +5.37%  (p=0.000 n=20+20)
	Marshal/Map/OneLarge   97.9MB/s ± 1%  92.6MB/s ± 1%   -5.48%  (p=0.000 n=20+18)
	Marshal/Map/ManySmall  45.8MB/s ± 1%  50.7MB/s ± 2%  +10.67%  (p=0.000 n=19+20)

	name                    old alloc/op   new alloc/op    delta
	Marshal/Map/ManyEmpty      320B ± 0%      320B ± 0%     ~     (all equal)
	Marshal/Map/OneLarge       256B ± 0%      256B ± 0%     ~     (all equal)
	Marshal/Map/ManySmall    1.24kB ± 0%    1.24kB ± 0%   -0.08%  (p=0.000 n=20+20)

	name                   old allocs/op  new allocs/op    delta
	Marshal/Map/ManyEmpty      1.00 ± 0%      1.00 ± 0%     ~     (all equal)
	Marshal/Map/OneLarge       3.00 ± 0%      3.00 ± 0%     ~     (all equal)
	Marshal/Map/ManySmall      55.0 ± 0%      55.0 ± 0%     ~     (all equal)

As can be seen, there is no change in the number of bytes allocated.
The benchmarks for small maps get faster since we no longer have
the overhead of getting/putting a reflect.MapIter to/from the sync.Pool.

Interestingly, the OneLarge benchmark gets slightly slower even though the
inner loop of makeMapArshaler is the same. Benchmarks show that the slowdown
occurs on k.SetIterKey(iter). One theory is that a pooled reflect.MapIter
was heap-allocated and happened to have better cache locality with the key.