Reject '\r' and '\n' in base32 and base64 encoded strings (#255)
RFC 4648, section 3.3, states:

    Implementations MUST reject the encoded data if it contains
    characters outside the base alphabet when interpreting base-encoded
    data, unless the specification referring to this document explicitly
    states otherwise.

We do not state otherwise. Thus, we should be rejecting '\r' and \'n'.
Also, other JSON specifications (e.g., RFC 7493) recommend the use of
base64 encoding, but do not call out any ignored characters.

Unfortunately, the "base32" and "base64" packages ignore '\r' and '\n'
by default, which is the wrong default. Since this is a new module,
we should opt into stricter and more correct behavior from the start.
It is easier to loosen this in the future, than making it more strict.

Note that we do not reject de-normalized inputs where the padding
contains non-zero bits. RFC 4648, section 3.5, states:

    In some environments, the alteration is critical and therefore
    decoders MAY chose to reject an encoding if the pad bits have not
    been set to zero.  The specification referring to this may mandate a
    specific behaviour.

It is not required that we reject non-zero padding, so we do not.