Strictly comply with RFC 3339 (#256)
RFC 3339 has a specified grammar for valid timestamps.

When marshaling, we should produce an error if the timestamp
cannot be represented in RFC 3339.
In particular, we should reject years and timezone offsets
that are out of range.

When unmarshaling, the Go "time" package unfortunately permits
timestamps that are not valid RFC 3339. Explicitly check for
and reject these cases.