Move testing functionality internal/jsontest (#295)
In preparation for splitting jsontext out as a separate package,
move over testing specific functionality into an internal package
that both json and jsontext will depend on.

In particular, we move all the testdata to internal/jsontest and
make it available in the Data variable. We use go:embed to embed
all the testdata directly into the binary.

We also coalesce the test case marking logic to this package.
A huge portion of the diff is just renaming of functions.