Use constructors for static tokens
WARNING: This commit includes breaking changes.

Instead of global variables, use function constructors
for static tokens such as Null, ObjectStart, ObjectEnd,
ArrayStart, and ArrayEnd. This provides a greater degree
of immutability for the API.

Note that static tokens for True and False are dropped
in favor of the pre-existing Bool constructor.
If necessary, we can already re-introduce True and False.