Reject JSON numbers with StringifyNumbers when unmarshaling
WARNING: This commit includes breaking changes.

Previously, if StringifyNumbers was specified,
we would permit unmarshaling of a JSON number into a Go number.
However, this is not how the v1 equivalent functionality operated.
In v1, once the `string` tag option was specified,
it would only unmarshal a JSON string containing a JSON number,
but would reject a JSON number.

It may be reasonable to provide a separate option that allows
the more flexible unmarshal behavior that this commit breaks,
but that can be a future change. For now, we are working towards
implementing v1 entirely in terms of v2.