Implement support for OmitZeroStructFields
By default, Marshal serializes Go structs with all fields
even if each field is the zero Go value.
This can lead to verbose output emitting fields for which it is
often indistinguishable from the zero value of the field itself.

The caller-specified OmitZeroStructFields is equivalent to
specifying the omitzero tag option on every Go struct field.
Static analysis of Go code used with JSON indicates that
about 50% of all struct fields have `omitempty` specified.
While `omitempty` is not the same as `omitzero`, it is probably
safer to provide a caller-scoped OmitZeroStructFields option since
doing so is more likely to produce output that is semantically
equivalent to if the option was not specified at all.

The exceptions to the above is if the field value contains a -0.0
or if it implements IsZero that treats certain values other than
the zero Go value as "zero".