Always update stack pointer
Previously, we used to populate the stack pointer only under
the semantics of AllowDuplicateNames(false).
The argument made at the time was that the stack pointer
becomes imprecise if there are duplicate names and
that someone might want to disable duplicate name checks
for the sake of performance in which case the logic
to maintain the name stack becomes an unnecessary burden
if the user does not care about ever retrieving the stack.

However, preserving the stack pointer is important for
providing good errors and v1 has functionally always
operated under AllowDuplicateNames(false) semantics.
Thus, in order to preserve good errors in v1,
always maintain the name stack needed to produce a JSON pointer.