Implement v1 in terms of v2
Remove all the core functionality of v1 and instead implement it
in terms of v2 with the appropriate options specified.

Many v1 test cases currently failure because the v1 emulation in v2
is not a sufficient reproduction of historial v1 behavior.

The list of known failures are stored in v1/failing.txt,
which allows such failing tests to be skipped.

As the v1 emulation in v2 improves, the failing.txt list
should eventually become empty.