Implement legacy semantics for merging
This includes a minor v2 change where it is valid to unmarshal
a JSON null into an encoding.TextMarshaler.
This is a reasonable change as all other types allow JSON null.

TestInterfaceSet was modernized and had more test cases added.
The behavior that it enforces is identical to what v1 does today.
This test change has been upstreamed as https://go.dev/cl/638256.