Implement legacy semantics for string escaping behavior (#85)
For doubly escaped strings, the v1 behavior was to apply
escaping for EscapeForHTML and EscapeForJS on the first pass,
rather than on the second pass. This does lead to unnecessarily
longer outputs since escaping on the first pass means that
the newly included '\\' characters have to be escaped again
in the second pass. Doubly escaped strings are already an
esoteric feature of v1, so we should preserve the semantic.

Also, when encoding the raw output of MarshalJSON,
the v1 behavior was to preserve any pre-existing escape sequences,
while still escaping any characters that might need escaping
per EscapeForHTML and EscapeForJS. Replicate this behavior.