Use AppendEncode and AppenDecode available in Go 1.22
As a minor behavior change, we still store the decoded data
even if the sizes mismatch. If the input is too large,
then the excess is truncated. If the input is too small,
then we zero out the remainder. This matches the behavior
of how array underflow and overflow are handled
for normal [N]T arrays that are not byte sequences.