Handle string option for time types similar to other types
WARNING: This commit includes breaking changes.

Previously, if StringifyNumbers was specified,
we would permit unmarshaling of a JSON number into a Go number.
However, this is not how the v1 equivalent functionality operated.
In v1, once the `string` tag option was specified,
it would only unmarshal a JSON string containing a JSON number,
but would reject a JSON number.

In #60, we made v2 more consistent with v1 for regular numeric types,
but forgot to do so for time.Duration as well.

Also, in #91, we no longer specify StringifyNumbers for map keys
and expect a type to check the JSON state machine to handle
object names specially. Do the same for time.Duration and time.Time.