Sort by UTF-8 instead of UTF-16
When operating under the Deterministic option,
sort names according to UTF-8 instead of UTF-16.

The practical difference is negligible,
but the JSONv2 discussion revealed that many
objected to sorted according to UTF-16,
which is what RFC 8785 specifies.

Given that Deterministic does not say we comply with RFC 8785,
we can choose any arbitrary ordering.