Rename MarshalJSONV2 and UnmarshalJSONV2 using To and From suffix
While the MarshalJSONV2 and UnmarshalJSONV2 methods are technically
the "v2" variants of the methods, we could use the To and From suffix
that conveys slightly more information.

For example, consider the following user-defined methods:

	func (v MyType) MarshalJSONTo(e *jsontext.Encoder, o json.Options) error
	func (v *MyType) UnmarshalJSONFrom(d *jsontext.Decoder, o json.Options) error

In English, this would naturally read as:

	* v will marshal JSON to e
	* v will unmarshal JSON from d

The use of the To and From suffix matches the pre-existing
io.WriterTo and io.ReaderFrom interfaces, which are optimized variants of
io.Writer and io.Reader.

Thus, we perform the following renames:

	* MarshalerV1 -> Marshaler
	* MarshalerV2 -> MarshalerTo
	* MarshalJSONV2 -> MarshalJSONTo
	* MarshalFuncV1 -> MarshalFunc
	* MarshalFuncV2 -> MarshalToFunc
	* UnmarshalerV1 -> Unmarshaler
	* UnmarshalerV2 -> UnmarshalerFrom
	* UnmarshalJSONV2 -> UnmarshalJSONFrom
	* UnmarshalFuncV1 -> UnmarshalFunc
	* UnmarshalFuncV2 -> UnmarshalFromFunc

Note that we keep around the older types and functions so that
we have at least 1 commit of dual compatibility.
These will be deleted in a subsequent commit.