Remove renamed Marshal and Unmarshal methods and functions
WARNING: This commit contains breaking changes.

Remove MarshalerV1, MarshalerV2, UnmarshalerV1, UnmarshalerV2,
MarshalFuncV1, MarshalFuncV2, UnmarshalFuncV1, UnmarshalFuncV2, and
support for MarshalJSONV1 and UnmarshalJSONV2 methods.

These have been replaced by
Marshaler, MarshalerTo, Unmarshaler, UnmarshalerFrom,
MarshalFunc, MarshalToFunc, UnmarshalFunc, UnmarshalFromFunc, and
new support for MarshalJSONTo and UnmarshalJSONFrom methods.

Of note, there will not necessarily be a static build failure
because of the renamed methods. It is the maintainer's responsibility
to verify that MarshalJSONV2 and UnmarshalJSONV2 methods are renamed
when updating a dependency on this module.

The prior commit contains support for both MarshalJSONV2 and MarshalJSONTo
as well as UnmarshalJSONV2 and UnmarshalJSONFrom.
Use that commit as a staging point to convert code to use the new API.