Keep more inline, but invalid fields (#127)
Some existing codebases include struct fields like the following:

    type T struct {
        F `json:",inline,omitempty"`
    }

This is odd since v1 "encoding/json" never supported an `inline` option.
This happened to still work in v1 since F is an embedded field.

However, v2 adds explicit support for the `inline` option,
but expects specifying it to be sensibly done.
In particular, there is no clear semantic what the combination
of both `inline` and `omitempty` means. The `inline` option means
that the entirety of all fields in F are promoted to type T
(effectively removing F itself as a serializable field).
However, the `omitempty` option means that F as a singular field
should not be serialized if empty. However, under `inline` semantics
F is no longer up for consideration as a singular field.
Thus, `omitempty` has no meaning when specified with `inline`.

Modify v2 to still report an error with this construction,
but at least still allow F to be treated as an inlined field.
In general, specifying `inline` with any other option means
that the other options are effectively ignored.
Note that under v1 semantics, struct errors are ignored.