Add back v1 implementation
This pulls in the v1 "encoding/json" implementation at commit 99dad5281660c4e644602e0c8790dd24b3eb45f3.

For all Marshal and Encoder.Encode calls, we execute both v1 and v2
and verify that it is byte-for-byte identical.

For all Unmarshal and Decoder.Decode calls, we execute both v1 and v2
if unmarshaling into a zero'd Go value and verify that
the values a deeply equal.